+++ docs/rules/img-redundant-alt.md
@@ -1,4 +1,4 @@
-# redundant-alt
+# img-redundant-alt
 
 Enforce img alt attribute does not contain the word image, picture, or photo.
 Screenreaders already announce `img` elements as an image. There is no need to use
@@ -13,7 +13,7 @@ This rule takes the following options:
 
 ## Passes
 
-```js
+```jsx harmony
 // passes when the `alt` does not contain redundant words
 <img src="foo" alt="nice"/>
 
@@ -27,7 +27,7 @@ This rule takes the following options:
 
 ## Fails
 
-```js
+```jsx harmony
 // fails when is a redundant alt message
 <img src="foo" alt="bar image foo"/>
