+++ src/rules/img-redundant-alt.js
