+++ src/rules/interactive-supports-focus.js
